feat: add overrides `num-columns` and `group-heading-level`
Before this change, `group-heading-level` defaulted to reusing the last
heading level found in the document. This is unlikely to always work for
users. This new parameter allows users to override the default.

Additionally, to give users some ability to style the output without
needing to override `user-print-glossary`, the new `num-columns`
parameter gives users the ability to control the number of columns
groups are printed in.